♻️ Extract PricingPageContent from PaywallModal for inline member page rendering
The member page previously opened PaywallModal as a fullscreen overlay hack.
Now it renders PricingPageContent directly as page content, while PaywallModal
remains a thin modal wrapper for other callers (e.g. TTS player upsell).
