+++ components/PricingPageContent.props.ts
@@ -1,0 +1,10 @@
+export interface PricingPageContentProps {
+  modelValue?: SubscriptionPlan
+  isProcessingSubscription?: boolean
+  trialPeriodDays?: number
+  mustCollectPaymentMethod?: boolean
+  utmCampaign?: string
+  utmMedium?: string
+  utmSource?: string
+  coupon?: string
+}
